+++ sources/dynamodb/schema.go
